LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-scope-dynamic-expected.txt


FAIL anchor-scope:all appearing dynamically assert_equals: expected "20px" but got "40px"
FAIL anchor-scope:--a appearing dynamically assert_equals: expected "20px" but got "40px"
PASS anchor-scope:--b appearing dynamically (--b never referenced)
FAIL anchor-scope:--a appearing dynamically scopes only --a assert_equals: expected "20px" but got "40px"
